+++ home_applience_service_app_front_end/package.json
@@ -14,6 +14,7 @@
     "heroicons-react": "^1.4.1",
     "lucide-react": "^0.375.0",
     "react": "^18.2.0",
+    "react-datepicker": "^6.9.0",
     "react-dom": "^18.2.0",
     "react-modal": "^3.16.1",
     "react-router-dom": "^6.23.0",
