+++ home_applience_service_app_front_end/src/App.js
@@ -1,11 +1,9 @@
 import MainRouter from "./service/MainRouter";
 
-function App() {
+export default function App() {
   return (
     <>
       <MainRouter />
     </>
   );
 }
-
-export default App;
