+++ home_applience_service_app_front_end/src/components/CustomerDashboard/Products.jsx
@@ -1,0 +1,1020 @@
+// import { useState } from "react";
+// import { Link } from "react-router-dom";
+// import { motion } from "framer-motion";
+// import Modal from "react-modal";
+// import axios from "axios";
+
+// const customStyles = {
+//   content: {
+//     top: "50%",
+//     left: "50%",
+//     right: "auto",
+//     bottom: "auto",
+//     width: "450px",
+//     marginRight: "-50%",
+//     transform: "translate(-50%, -50%)",
+//     maxWidth: "500px",
+//   },
+// };
+
+// const Products = () => {
+//   const [modalIsOpen, setIsOpen] = useState(false);
+//   const [selectedProduct, setSelectedProduct] = useState(null);
+//   const [formData, setFormData] = useState({
+//     applianceBrand: "",
+//     yearOfManufacturing: "",
+//     productType: "",
+//     serialNo: "",
+//     warrantyStatus: "",
+//     email: "",
+//   });
+
+//   const applianceBrandOptions = [
+//     "LG",
+//     "SONY",
+//     "BOSCH",
+//     "SAMSUNG",
+//     "TOSHIBA",
+//     "TPLINK",
+//     "DLINK",
+//   ];
+//   const yearOptions = [
+//     "2024-01-01",
+//     "2023-01-01",
+//     "2022-01-01",
+//     "2021-01-01",
+//     "2020-01-01",
+//   ];
+//   const productTypeOptions = [
+//     "FAN",
+//     "AIRCONDITIONER",
+//     "WASHINGMACHINE",
+//     "WIRELESSROUTER",
+//     "COFFEEDISPENSER",
+//     "REFRIGERATOR",
+//   ];
+//   const warrantyOptions = ["UNDERWARRENTY", "OUTOFWARRENTY"];
+
+//   const products = [
+//     {
+//       id: 1,
+//       name: "Fan",
+//       to: "#",
+//       imageSrc:
+//         "https://images.unsplash.com/photo-1564510182791-29645da7fac4?w=400&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8NHx8RmFufGVufDB8fDB8fHww",
+//       imageAlt: "Fan",
+//     },
+//     {
+//       id: 2,
+//       name: "Air Conditioner",
+//       to: "#",
+//       imageSrc:
+//         "https://images.unsplash.com/photo-1436473849883-bb3464c23e93?q=80&w=1470&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+//       imageAlt: "AC",
+//     },
+//     {
+//       id: 3,
+//       name: "Washing Machine",
+//       to: "#",
+//       imageSrc:
+//         "https://images.unsplash.com/photo-1626806787461-102c1bfaaea1?w=400&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8M3x8V2FzaGluZyUyMG1hY2hpbmV8ZW58MHx8MHx8fDA%3D",
+//       imageAlt: "Washing Machine",
+//     },
+//     {
+//       id: 4,
+//       name: "Wireless Routers",
+//       to: "#",
+//       imageSrc:
+//         "https://plus.unsplash.com/premium_photo-1675024346167-e5fd43181f16?w=400&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8MTd8fHdpcmVsZXNzJTIwcm91dGVyfGVufDB8fDB8fHww",
+//       imageAlt: "Wireless router",
+//     },
+//     {
+//       id: 5,
+//       name: "Coffee Dispenser",
+//       to: "#",
+//       imageSrc:
+//         "https://plus.unsplash.com/premium_photo-1661281243564-5acf9a1797ed?w=400&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8NXx8Y29mZmVlJTIwbWFjaGluZXxlbnwwfHwwfHx8MA%3D%3D",
+//       imageAlt: "Coffee Dispenser",
+//     },
+//     {
+//       id: 6,
+//       name: "Refrigerator",
+//       to: "#",
+//       imageSrc:
+//         "https://media.istockphoto.com/id/990894662/photo/modern-fridge-in-front-of-white-wall.webp?b=1&s=170667a&w=0&k=20&c=WMEUix4zvDRc0baQ2i-E1IxjpS6lXw_vWgiIGeKurKk=",
+//       imageAlt: "Refrigerator",
+//     },
+//   ];
+
+//   // openModal for appliance
+
+//   const openModal = (product) => {
+//     setSelectedProduct(product);
+//     setFormData({
+//       ...formData,
+//       productType: product.name.toUpperCase().replace(/\s/g, ""),
+//     });
+//     setIsOpen(true);
+//   };
+
+//   const closeModal = () => {
+//     setIsOpen(false);
+//   };
+
+//   const handleInputChange = (e) => {
+//     const { name, value } = e.target;
+//     setFormData({ ...formData, [name]: value });
+//   };
+
+//   const handleSubmit = async (e) => {
+//     e.preventDefault();
+//     try {
+//       const response = await axios.post(
+//         "http://localhost:8080/appliance/save",
+//         formData
+//       );
+//       alert(
+//         "Appointment has been successfully registered, You will get a call from our technician team shortly",
+//         response.data
+//       );
+//       closeModal();
+//     } catch (error) {
+//       console.error("Error submitting form:", error);
+//     }
+//   };
+
+//   // openModal for Service Request
+
+//   const openModalService = (product) => {
+//     setSelectedProduct(product);
+//     setFormData({
+//       ...formData,
+//       productType: product.name.toUpperCase().replace(/\s/g, ""),
+//     });
+//     setIsOpen(true);
+//   };
+
+//   const closeModalService = () => {
+//     setIsOpen(false);
+//   };
+
+//   const handleInputChangeService = (e) => {
+//     const { name, value } = e.target;
+//     setFormData({ ...formData, [name]: value });
+//   };
+
+//   const handleSubmitService = async (e) => {
+//     e.preventDefault();
+//     try {
+//       const response = await axios.post(
+//         "http://localhost:8080/appliance/save",
+//         formData
+//       );
+//       alert(
+//         "Appointment has been successfully registered, You will get a call from our technician team shortly",
+//         response.data
+//       );
+//       closeModalService();
+//     } catch (error) {
+//       console.error("Error submitting form:", error);
+//     }
+//   };
+//   return (
+//     <>
+//       <motion.div
+//         initial={{ opacity: 0 }}
+//         animate={{ opacity: 1 }}
+//         exit={{ opacity: 0 }}
+//         transition={{ duration: 1.5 }}
+//       >
+//         <div className="bg-white">
+//           <div className="max-w-7xl mx-auto py-16 px-4 overflow-hidden sm:py-24 sm:px-6 lg:px-8">
+//             <div className="grid grid-cols-1 gap-y-10 gap-x-6 sm:grid-cols-2 lg:grid-cols-3 lg:gap-x-8">
+//               {products.map((product) => (
+//                 <div key={product.id}>
+//                   <Link to={product.href} className="group text-sm">
+//                     <div className="w-full aspect-w-1 aspect-h-1 rounded-lg overflow-hidden bg-gray-100 group-hover:opacity-75">
+//                       <img
+//                         src={product.imageSrc}
+//                         alt={product.imageAlt}
+//                         className="w-full h-full object-center object-cover"
+//                       />
+//                     </div>
+//                     <h3 className="mt-4 font-medium text-gray-900">
+//                       {product.name}
+//                     </h3>
+//                     <p className="text-gray-500 italic">
+//                       {product.availability}
+//                     </p>
+//                     <p className="mt-2 font-medium text-gray-900">
+//                       {product.price}
+//                     </p>
+//                   </Link>
+//                   <button
+//                     type="button"
+//                     className="mt-5 bg-transparent hover:bg-indigo-500 text-indigo-700 font-semibold hover:text-white py-1 px-3 border border-indigo-500 hover:border-transparent rounded-2xl"
+//                     onClick={() => openModal(product)}
+//                   >
+//                     Book Appointment
+//                   </button>
+//                   <button
+//                     type="button"
+//                     className="mt-5 mx-4 bg-transparent hover:bg-indigo-500 text-indigo-700 font-semibold hover:text-white py-1 px-3 border border-indigo-500 hover:border-transparent rounded-2xl"
+//                     onClick={() => openModalService(product)}
+//                   >
+//                     Service Request
+//                   </button>
+//                 </div>
+//               ))}
+//             </div>
+//           </div>
+//         </div>
+//       </motion.div>
+//       <Modal
+//         isOpen={modalIsOpen}
+//         onRequestClose={closeModal}
+//         style={customStyles}
+//         contentLabel="Book Appointment Modal"
+//       >
+//         <h2 className="text-xl font-bold mb-4">
+//           Book Appointment for {selectedProduct && selectedProduct.name}
+//         </h2>
+//         <form onSubmit={handleSubmit}>
+//           <div className="mb-4">
+//             <label htmlFor="email" className="block mb-1">
+//               Email
+//             </label>
+//             <input
+//               type="email"
+//               id="email"
+//               name="email"
+//               value={formData.email}
+//               onChange={handleInputChange}
+//               className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+//               required
+//             />
+//           </div>
+//           <div className="mb-4">
+//             <label htmlFor="applianceBrand" className="block mb-1">
+//               Appliance Brand
+//             </label>
+//             <select
+//               id="applianceBrand"
+//               name="applianceBrand"
+//               value={formData.applianceBrand}
+//               onChange={handleInputChange}
+//               className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+//               required
+//             >
+//               <option value="">Select Brand</option>
+//               {applianceBrandOptions.map((option) => (
+//                 <option key={option} value={option}>
+//                   {option}
+//                 </option>
+//               ))}
+//             </select>
+//           </div>
+//           <div className="mb-4">
+//             <label htmlFor="yearOfManufacturing" className="block mb-1">
+//               Year of Manufacturing
+//             </label>
+//             <select
+//               id="yearOfManufacturing"
+//               name="yearOfManufacturing"
+//               value={formData.yearOfManufacturing}
+//               onChange={handleInputChange}
+//               className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+//               required
+//             >
+//               <option value="">Select Year</option>
+//               {yearOptions.map((option) => (
+//                 <option key={option} value={option}>
+//                   {option}
+//                 </option>
+//               ))}
+//             </select>
+//           </div>
+//           <div className="mb-4">
+//             <label htmlFor="productType" className="block mb-1">
+//               Product Type
+//             </label>
+//             <select
+//               id="productType"
+//               name="productType"
+//               value={formData.productType}
+//               onChange={handleInputChange}
+//               className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+//               required
+//             >
+//               <option value="">Select Product Type</option>
+//               {productTypeOptions.map((option) => (
+//                 <option key={option} value={option}>
+//                   {option}
+//                 </option>
+//               ))}
+//             </select>
+//           </div>
+//           <div className="mb-4">
+//             <label htmlFor="serialNo" className="block mb-1">
+//               Serial No
+//             </label>
+//             <input
+//               type="text"
+//               id="serialNo"
+//               name="serialNo"
+//               value={formData.serialNo}
+//               onChange={handleInputChange}
+//               className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+//               required
+//             />
+//           </div>
+//           <div className="mb-4">
+//             <label htmlFor="warrantyStatus" className="block mb-1">
+//               Warranty Status
+//             </label>
+//             <select
+//               id="warrentyStatus"
+//               name="warrentyStatus"
+//               value={formData.warrentyStatus}
+//               onChange={handleInputChange}
+//               className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+//               required
+//             >
+//               <option value="">Select Warranty Status</option>
+//               {warrantyOptions.map((option) => (
+//                 <option key={option} value={option}>
+//                   {option}
+//                 </option>
+//               ))}
+//             </select>
+//           </div>
+//           <div className="text-center">
+//             <button
+//               type="submit"
+//               className="bg-indigo-500 text-white font-semibold py-2 px-6 rounded-md hover:bg-indigo-600 focus:outline-none focus:bg-indigo-600"
+//             >
+//               Submit
+//             </button>
+//             <button
+//               type="button"
+//               onClick={closeModal}
+//               className="ml-2 bg-gray-300 text-gray-700 font-semibold py-2 px-6 rounded-md hover:bg-gray-400 focus:outline-none focus:bg-gray-400"
+//             >
+//               Cancel
+//             </button>
+//           </div>
+//         </form>
+//       </Modal>
+//       // Modal for Service Request
+//       <Modal
+//         isOpen={modalIsOpen}
+//         onRequestClose={closeModalService}
+//         style={customStyles}
+//         contentLabel="Service Request Modal"
+//       >
+//         <h2 className="text-xl font-bold mb-4">
+//           Raise a Service Request for {selectedProduct && selectedProduct.name}
+//         </h2>
+//         <form onSubmit={handleSubmitService}>
+//           <div className="mb-4">
+//             <label htmlFor="email" className="block mb-1">
+//               Email
+//             </label>
+//             <input
+//               type="email"
+//               id="email"
+//               name="email"
+//               value={formData.email}
+//               onChange={handleInputChangeService}
+//               className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+//               required
+//             />
+//           </div>
+//           <div className="mb-4">
+//             <label htmlFor="applianceBrand" className="block mb-1">
+//               Created On
+//             </label>
+//             <select
+//               id="applianceBrand"
+//               name="applianceBrand"
+//               value={formData.applianceBrand}
+//               onChange={handleInputChangeService}
+//               className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+//               required
+//             >
+//               <option value="">Select Brand</option>
+//               {applianceBrandOptions.map((option) => (
+//                 <option key={option} value={option}>
+//                   {option}
+//                 </option>
+//               ))}
+//             </select>
+//           </div>
+//           <div className="mb-4">
+//             <label htmlFor="yearOfManufacturing" className="block mb-1">
+//               Updated On
+//             </label>
+//             <select
+//               id="yearOfManufacturing"
+//               name="yearOfManufacturing"
+//               value={formData.yearOfManufacturing}
+//               onChange={handleInputChangeService}
+//               className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+//               required
+//             >
+//               <option value="">Select Year</option>
+//               {yearOptions.map((option) => (
+//                 <option key={option} value={option}>
+//                   {option}
+//                 </option>
+//               ))}
+//             </select>
+//           </div>
+//           <div className="mb-4">
+//             <label htmlFor="productType" className="block mb-1">
+//               Appointment Date
+//             </label>
+//             <select
+//               id="productType"
+//               name="productType"
+//               value={formData.productType}
+//               onChange={handleInputChangeService}
+//               className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+//               required
+//             >
+//               <option value="">Select Product Type</option>
+//               {productTypeOptions.map((option) => (
+//                 <option key={option} value={option}>
+//                   {option}
+//                 </option>
+//               ))}
+//             </select>
+//           </div>
+//           <div className="mb-4">
+//             <label htmlFor="serialNo" className="block mb-1">
+//               Service Status
+//             </label>
+//             <input
+//               type="text"
+//               id="serialNo"
+//               name="serialNo"
+//               value={formData.serialNo}
+//               onChange={handleInputChangeService}
+//               className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+//               required
+//             />
+//           </div>
+//           <div className="mb-4">
+//             <label htmlFor="warrantyStatus" className="block mb-1">
+//               Comments
+//             </label>
+//             <select
+//               id="warrentyStatus"
+//               name="warrentyStatus"
+//               value={formData.warrentyStatus}
+//               onChange={handleInputChangeService}
+//               className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+//               required
+//             >
+//               <option value="">Select Warranty Status</option>
+//               {warrantyOptions.map((option) => (
+//                 <option key={option} value={option}>
+//                   {option}
+//                 </option>
+//               ))}
+//             </select>
+//           </div>
+//           <div className="text-center">
+//             <button
+//               type="submit"
+//               className="bg-indigo-500 text-white font-semibold py-2 px-6 rounded-md hover:bg-indigo-600 focus:outline-none focus:bg-indigo-600"
+//             >
+//               Submit
+//             </button>
+//             <button
+//               type="button"
+//               onClick={closeModalService}
+//               className="ml-2 bg-gray-300 text-gray-700 font-semibold py-2 px-6 rounded-md hover:bg-gray-400 focus:outline-none focus:bg-gray-400"
+//             >
+//               Cancel
+//             </button>
+//           </div>
+//         </form>
+//       </Modal>
+//     </>
+//   );
+// };
+
+// export default Products;
+
+import { useState } from "react";
+import { Link } from "react-router-dom";
+import { motion } from "framer-motion";
+import Modal from "react-modal";
+import axios from "axios";
+
+const customStyles = {
+  content: {
+    top: "50%",
+    left: "50%",
+    right: "auto",
+    bottom: "auto",
+    width: "450px",
+    marginRight: "-50%",
+    transform: "translate(-50%, -50%)",
+    maxWidth: "500px",
+  },
+};
+
+const Products = () => {
+  const [modalIsOpen, setIsOpen] = useState(false);
+  const [selectedProduct, setSelectedProduct] = useState(null);
+  const [formData, setFormData] = useState({
+    applianceBrand: "",
+    yearOfManufacturing: "",
+    productType: "",
+    serialNo: "",
+    warrantyStatus: "",
+    email: "",
+    createdOn: "",
+    updatedOn: "",
+    appointmentDate: "",
+  });
+
+  const applianceBrandOptions = [
+    "LG",
+    "SONY",
+    "BOSCH",
+    "SAMSUNG",
+    "TOSHIBA",
+    "TPLINK",
+    "DLINK",
+  ];
+  const yearOptions = [
+    "2024-01-01",
+    "2023-01-01",
+    "2022-01-01",
+    "2021-01-01",
+    "2020-01-01",
+  ];
+  const productTypeOptions = [
+    "FAN",
+    "AIRCONDITIONER",
+    "WASHINGMACHINE",
+    "WIRELESSROUTER",
+    "COFFEEDISPENSER",
+    "REFRIGERATOR",
+  ];
+  const warrantyOptions = ["UNDERWARRENTY", "OUTOFWARRENTY"];
+
+  const products = [
+    {
+      id: 1,
+      name: "Fan",
+      to: "#",
+      imageSrc:
+        "https://images.unsplash.com/photo-1564510182791-29645da7fac4?w=400&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8NHx8RmFufGVufDB8fDB8fHww",
+      imageAlt: "Fan",
+    },
+    {
+      id: 2,
+      name: "Air Conditioner",
+      to: "#",
+      imageSrc:
+        "https://images.unsplash.com/photo-1436473849883-bb3464c23e93?q=80&w=1470&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D",
+      imageAlt: "AC",
+    },
+    {
+      id: 3,
+      name: "Washing Machine",
+      to: "#",
+      imageSrc:
+        "https://images.unsplash.com/photo-1626806787461-102c1bfaaea1?w=400&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8M3x8V2FzaGluZyUyMG1hY2hpbmV8ZW58MHx8MHx8fDA%3D",
+      imageAlt: "Washing Machine",
+    },
+    {
+      id: 4,
+      name: "Wireless Routers",
+      to: "#",
+      imageSrc:
+        "https://plus.unsplash.com/premium_photo-1675024346167-e5fd43181f16?w=400&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8MTd8fHdpcmVsZXNzJTIwcm91dGVyfGVufDB8fDB8fHww",
+      imageAlt: "Wireless router",
+    },
+    {
+      id: 5,
+      name: "Coffee Dispenser",
+      to: "#",
+      imageSrc:
+        "https://plus.unsplash.com/premium_photo-1661281243564-5acf9a1797ed?w=400&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8NXx8Y29mZmVlJTIwbWFjaGluZXxlbnwwfHwwfHx8MA%3D%3D",
+      imageAlt: "Coffee Dispenser",
+    },
+    {
+      id: 6,
+      name: "Refrigerator",
+      to: "#",
+      imageSrc:
+        "https://media.istockphoto.com/id/990894662/photo/modern-fridge-in-front-of-white-wall.webp?b=1&s=170667a&w=0&k=20&c=WMEUix4zvDRc0baQ2i-E1IxjpS6lXw_vWgiIGeKurKk=",
+      imageAlt: "Refrigerator",
+    },
+  ];
+
+  // openModal for appliance
+
+  const openModal = (product) => {
+    setSelectedProduct(product);
+    setFormData({
+      ...formData,
+      productType: product.name.toUpperCase().replace(/\s/g, ""),
+    });
+    setIsOpen(true);
+  };
+
+  const closeModal = () => {
+    setIsOpen(false);
+  };
+
+  const handleInputChange = (e) => {
+    const { name, value } = e.target;
+    setFormData({ ...formData, [name]: value });
+  };
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    try {
+      const response = await axios.post(
+        "http://localhost:8080/appliance/save",
+        formData
+      );
+      alert(
+        "Appointment has been successfully registered, You will get a call from our technician team shortly",
+        response.data
+      );
+      closeModal();
+    } catch (error) {
+      console.error("Error submitting form:", error);
+    }
+  };
+
+  // openModal for Service Request
+
+  const openModalService = (product) => {
+    setSelectedProduct(product);
+    setFormData({
+      ...formData,
+      productType: product.name.toUpperCase().replace(/\s/g, ""),
+    });
+    setIsOpen(true);
+  };
+
+  const closeModalService = () => {
+    setIsOpen(false);
+  };
+
+  const handleInputChangeService = (e) => {
+    const { name, value } = e.target;
+    setFormData({ ...formData, [name]: value });
+  };
+
+  const handleSubmitService = async (e) => {
+    e.preventDefault();
+    try {
+      const response = await axios.post(
+        "http://localhost:8080/appliance/save",
+        formData
+      );
+      alert(
+        "Appointment has been successfully registered, You will get a call from our technician team shortly",
+        response.data
+      );
+      closeModalService();
+    } catch (error) {
+      console.error("Error submitting form:", error);
+    }
+  };
+  return (
+    <>
+      <motion.div
+        initial={{ opacity: 0 }}
+        animate={{ opacity: 1 }}
+        exit={{ opacity: 0 }}
+        transition={{ duration: 1.5 }}
+      >
+        <div className="bg-white">
+          <div className="max-w-7xl mx-auto py-16 px-4 overflow-hidden sm:py-24 sm:px-6 lg:px-8">
+            <div className="grid grid-cols-1 gap-y-10 gap-x-6 sm:grid-cols-2 lg:grid-cols-3 lg:gap-x-8">
+              {products.map((product) => (
+                <div key={product.id}>
+                  <Link to={product.href} className="group text-sm">
+                    <div className="w-full aspect-w-1 aspect-h-1 rounded-lg overflow-hidden bg-gray-100 group-hover:opacity-75">
+                      <img
+                        src={product.imageSrc}
+                        alt={product.imageAlt}
+                        className="w-full h-full object-center object-cover"
+                      />
+                    </div>
+                    <h3 className="mt-4 font-medium text-gray-900">
+                      {product.name}
+                    </h3>
+                    <p className="text-gray-500 italic">
+                      {product.availability}
+                    </p>
+                    <p className="mt-2 font-medium text-gray-900">
+                      {product.price}
+                    </p>
+                  </Link>
+                  <button
+                    type="button"
+                    className="mt-5 bg-transparent hover:bg-indigo-500 text-indigo-700 font-semibold hover:text-white py-1 px-3 border border-indigo-500 hover:border-transparent rounded-2xl"
+                    onClick={() => openModal(product)}
+                  >
+                    Book Appointment
+                  </button>
+                  <button
+                    type="button"
+                    className="mt-5 mx-4 bg-transparent hover:bg-indigo-500 text-indigo-700 font-semibold hover:text-white py-1 px-3 border border-indigo-500 hover:border-transparent rounded-2xl"
+                    onClick={() => openModalService(product)}
+                  >
+                    Service Request
+                  </button>
+                </div>
+              ))}
+            </div>
+          </div>
+        </div>
+      </motion.div>
+      <Modal
+        isOpen={modalIsOpen}
+        onRequestClose={closeModal}
+        style={customStyles}
+        contentLabel="Book Appointment Modal"
+      >
+        <h2 className="text-xl font-bold mb-4">
+          Book Appointment for {selectedProduct && selectedProduct.name}
+        </h2>
+        <form onSubmit={handleSubmit}>
+          <div className="mb-4">
+            <label htmlFor="email" className="block mb-1">
+              Email
+            </label>
+            <input
+              type="email"
+              id="email"
+              name="email"
+              value={formData.email}
+              onChange={handleInputChange}
+              className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+              required
+            />
+          </div>
+          <div className="mb-4">
+            <label htmlFor="applianceBrand" className="block mb-1">
+              Appliance Brand
+            </label>
+            <select
+              id="applianceBrand"
+              name="applianceBrand"
+              value={formData.applianceBrand}
+              onChange={handleInputChange}
+              className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+              required
+            >
+              <option value="">Select Brand</option>
+              {applianceBrandOptions.map((option) => (
+                <option key={option} value={option}>
+                  {option}
+                </option>
+              ))}
+            </select>
+          </div>
+          <div className="mb-4">
+            <label htmlFor="yearOfManufacturing" className="block mb-1">
+              Year of Manufacturing
+            </label>
+            <select
+              id="yearOfManufacturing"
+              name="yearOfManufacturing"
+              value={formData.yearOfManufacturing}
+              onChange={handleInputChange}
+              className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+              required
+            >
+              <option value="">Select Year</option>
+              {yearOptions.map((option) => (
+                <option key={option} value={option}>
+                  {option}
+                </option>
+              ))}
+            </select>
+          </div>
+          <div className="mb-4">
+            <label htmlFor="productType" className="block mb-1">
+              Product Type
+            </label>
+            <select
+              id="productType"
+              name="productType"
+              value={formData.productType}
+              onChange={handleInputChange}
+              className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+              required
+            >
+              <option value="">Select Product Type</option>
+              {productTypeOptions.map((option) => (
+                <option key={option} value={option}>
+                  {option}
+                </option>
+              ))}
+            </select>
+          </div>
+          <div className="mb-4">
+            <label htmlFor="serialNo" className="block mb-1">
+              Serial No
+            </label>
+            <input
+              type="text"
+              id="serialNo"
+              name="serialNo"
+              value={formData.serialNo}
+              onChange={handleInputChange}
+              className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+              required
+            />
+          </div>
+          <div className="mb-4">
+            <label htmlFor="warrantyStatus" className="block mb-1">
+              Warranty Status
+            </label>
+            <select
+              id="warrentyStatus"
+              name="warrentyStatus"
+              value={formData.warrentyStatus}
+              onChange={handleInputChange}
+              className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+              required
+            >
+              <option value="">Select Warranty Status</option>
+              {warrantyOptions.map((option) => (
+                <option key={option} value={option}>
+                  {option}
+                </option>
+              ))}
+            </select>
+          </div>
+          <div className="text-center">
+            <button
+              type="submit"
+              className="bg-indigo-500 text-white font-semibold py-2 px-6 rounded-md hover:bg-indigo-600 focus:outline-none focus:bg-indigo-600"
+            >
+              Submit
+            </button>
+            <button
+              type="button"
+              onClick={closeModal}
+              className="ml-2 bg-gray-300 text-gray-700 font-semibold py-2 px-6 rounded-md hover:bg-gray-400 focus:outline-none focus:bg-gray-400"
+            >
+              Cancel
+            </button>
+          </div>
+        </form>
+      </Modal>
+      {/* Modal for Service Request */}
+      <Modal
+        isOpen={modalIsOpen}
+        onRequestClose={closeModalService}
+        style={customStyles}
+        contentLabel="Service Request Modal"
+      >
+        <h2 className="text-xl font-bold mb-4">
+          Raise a Service Request for {selectedProduct && selectedProduct.name}
+        </h2>
+        <form onSubmit={handleSubmitService}>
+          <div className="mb-4">
+            <label htmlFor="email" className="block mb-1">
+              Email
+            </label>
+            <input
+              type="email"
+              id="email"
+              name="email"
+              value={formData.email}
+              onChange={handleInputChangeService}
+              className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+              required
+            />
+          </div>
+          <div className="mb-4">
+            <label htmlFor="applianceBrand" className="block mb-1">
+              Created On
+            </label>
+            <select
+              id="applianceBrand"
+              name="applianceBrand"
+              value={formData.applianceBrand}
+              onChange={handleInputChangeService}
+              className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+              required
+            >
+              <option value="">Select Brand</option>
+              {applianceBrandOptions.map((option) => (
+                <option key={option} value={option}>
+                  {option}
+                </option>
+              ))}
+            </select>
+          </div>
+          <div className="mb-4">
+            <label htmlFor="yearOfManufacturing" className="block mb-1">
+              Updated On
+            </label>
+            <select
+              id="yearOfManufacturing"
+              name="yearOfManufacturing"
+              value={formData.yearOfManufacturing}
+              onChange={handleInputChangeService}
+              className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+              required
+            >
+              <option value="">Select Year</option>
+              {yearOptions.map((option) => (
+                <option key={option} value={option}>
+                  {option}
+                </option>
+              ))}
+            </select>
+          </div>
+          <div className="mb-4">
+            <label htmlFor="productType" className="block mb-1">
+              Appointment Date
+            </label>
+            <select
+              id="productType"
+              name="productType"
+              value={formData.productType}
+              onChange={handleInputChangeService}
+              className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+              required
+            >
+              <option value="">Select Product Type</option>
+              {productTypeOptions.map((option) => (
+                <option key={option} value={option}>
+                  {option}
+                </option>
+              ))}
+            </select>
+          </div>
+          <div className="mb-4">
+            <label htmlFor="serialNo" className="block mb-1">
+              Service Status
+            </label>
+            <input
+              type="text"
+              id="serialNo"
+              name="serialNo"
+              value={formData.serialNo}
+              onChange={handleInputChangeService}
+              className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+              required
+            />
+          </div>
+          <div className="mb-4">
+            <label htmlFor="warrantyStatus" className="block mb-1">
+              Comments
+            </label>
+            <select
+              id="warrentyStatus"
+              name="warrentyStatus"
+              value={formData.warrentyStatus}
+              onChange={handleInputChangeService}
+              className="w-full px-4 py-2 border rounded-md focus:outline-none focus:border-indigo-500"
+              required
+            >
+              <option value="">Select Warranty Status</option>
+              {warrantyOptions.map((option) => (
+                <option key={option} value={option}>
+                  {option}
+                </option>
+              ))}
+            </select>
+          </div>
+          <div className="text-center">
+            <button
+              type="submit"
+              className="bg-indigo-500 text-white font-semibold py-2 px-6 rounded-md hover:bg-indigo-600 focus:outline-none focus:bg-indigo-600"
+            >
+              Submit
+            </button>
+            <button
+              type="button"
+              onClick={closeModalService}
+              className="ml-2 bg-gray-300 text-gray-700 font-semibold py-2 px-6 rounded-md hover:bg-gray-400 focus:outline-none focus:bg-gray-400"
+            >
+              Cancel
+            </button>
+          </div>
+        </form>
+      </Modal>
+    </>
+  );
+};
+
+export default Products;
